+++ src/typescript/src/domain/repositories/LampRepository.ts
